# openapi.yaml
openapi: 3.1.1
info:
  title: Engineering Thesis API
  version: 1.0.0
servers:
  - url: http://localhost:8080
components:
  securitySchemes:
    JwtBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateEnterpriseRequest:
      type: object
      required: [name, description, address, longitude, latitude]
      properties:
        name:
          type: string
          example: The Best Barbershop
        description:
          type: string
          example: The new best barbershop in town.
        address:
          type: string
          example: ul. Nowowiejska 15/19, Warsaw, Poland
        longitude:
          type: number
          format: double
          example: 12.3248
        latitude:
          type: number
          format: double
          example: -67.2247
        logoFile:
          type: string
          format: binary
        backgroundPhotoFile:
          type: string
          format: binary
    CreateEnterpriseResponse:
      type: object
      required: [enterpriseId]
      properties:
        enterpriseId:
          type: integer
          format: int64
          example: 345
    GetEnterpriseResponse:
      type: object
      required: [name, description]
      properties:
        name:
          type: string
          example: The Best Barbershop
        description:
          type: string
          example: The new best barbershop in town.
        location:
          $ref: "#/components/schemas/Location"
    CreateEnterpriseEmployeeRequest:
      type: object
      required:
        - firstName
        - lastName
        - username
        - password
      properties:
        firstName:
          type: string
          example: Jan
        lastName:
          type: string
          example: Kowalski
        username:
          type: string
          example: jan.kowalski.1
        password:
          type: string
          format: password
          example: "5=2VPIzYHSIc0l4d1o&es]plB"
    CreateEnterpriseEmployeeResponse:
      type: object
      required:
        - user
      properties:
        user:
          type: object
          required:
            - firstName
            - lastName
            - username
          properties:
            firstName:
              type: string
              example: Jan
            lastName:
              type: string
              example: Kowalski
            username:
              type: string
              example: jan.kowalski.1
    GetEnterpriseEmployeesResponse:
      type: array
      items:
        $ref: "#/components/schemas/GetEnterpriseEmployeesResponseItem"
    GetEnterpriseEmployeesResponseItem:
      type: object
      required: [username, firstName, lastName]
      properties:
        username:
          type: string
          example: jan.kowalski.1
        firstName:
          type: string
          example: Jan
        lastName:
          type: string
          example: Kowalski
...
    GetEnterpriseServiceCustomServiceResponse:
      type: object
      required:
        [
          takesCustomAppointments,
          serviceId,
          name,
          description,
          location,
          timezone,
          maxDistanceKm,
          cathegory,
          price,
          currencyIso,
        ]
      properties:
        takesCustomAppointments:
          type: boolean
          default: true
          readOnly: true
        serviceId:
          type: integer
          format: int64
          example: 1234
        name:
          type: string
          example: Basic haircut
        description:
          type: string
          example: Male or female basic haircut
        location:
          $ref: "#/components/schemas/Location"
        timezone:
          $ref: "#/components/schemas/IanaTimeZone"
        maxDistanceKm:
          type: number
          format: double
        cathegory:
          type: string
        price:
          type: number
        currencyIso:
          type: string
    GetEnterpriseServiceNonCustomServiceResponse:
      type: object
      required:
        [
          takesCustomAppointments,
          serviceId,
          name,
          description,
          location,
          timezone,
          cathegory,
          price,
          currencyIso,
        ]
      properties:
        takesCustomAppointments:
          type: boolean
          default: true
          readOnly: true
        serviceId:
          type: integer
          format: int64
          example: 1234
        name:
          type: string
          example: Basic haircut
        description:
          type: string
          example: Male or female basic haircut
        location:
          $ref: "#/components/schemas/Location"
        timezone:
          $ref: "#/components/schemas/IanaTimeZone"
        cathegory:
          type: string
        price:
          type: number
        currencyIso:
          type: string
    GetEnterpriseServicesResponse:
      type: array
      items:
        oneOf:
          - $ref: "#/components/schemas/GetEnterpriseServiceNonCustomServiceResponse"
          - $ref: "#/components/schemas/GetEnterpriseServiceCustomServiceResponse"
    PatchEnterpriseRequest:
      type: object
      required: [name, description, location, timeZone, takesCustomAppointments]
      properties:
        name:
          type: string
          example: The Best Barbershop
        description:
          type: string
          example: The new best barbershop in town.
        location:
          $ref: "#/components/schemas/Location"
        timeZone:
          type: string
        takesCustomAppointments:
          type: boolean
        logoFile:
          type: string
          format: binary
        backgroundPhotoFile:
          type: string
          format: binary
    CreateNoCustomAppointmentsEnterpriseServiceRequest:
      type: object
      required:
        [
          name,
          description,
          location,
          timeZone,
          cathegory,
          currency,
          price,
          slots,
        ]
      properties:
        name:
          type: string
          example: "Basic Haircut"
        description:
          type: string
          example: "Male or female haircut"
        location:
          $ref: "#/components/schemas/Location"
        timeZone:
          $ref: "#/components/schemas/IanaTimeZone"
        cathegory:
          type: string
        price:
          type: number
          example: 29.99
        currency:
          type: string
          example: "PLN"
          description: currency ISO code
        slots:
          type: array
          items:
            $ref: "#/components/schemas/Slot"
    CreateCustomAppointmentsEnterpriseServiceRequest:
      type: object
      required:
        [
          name,
          description,
          location,
          timeZone,
          maxDistanceKm,
          cathegory,
          price,
          currency,
          timeWindows,
        ]
      properties:
        name:
          type: string
          example: "Basic Haircut"
        description:
          type: string
          example: "Male or female haircut"
        location:
          $ref: "#/components/schemas/Location"
        timeZone:
          $ref: "#/components/schemas/IanaTimeZone"
        maxDistanceKm:
          type: number
          format: double
        cathegory:
          type: string
        price:
          type: number
          example: 29.99
        currency:
          type: string
          example: "PLN"
          description: currency ISO code
        timeWindows:
          type: array
          items:
            $ref: "#/components/schemas/TimeWindow"
    DayOfWeek:
      type: string
      enum: [MONDAY, TUESDAY, WEDNESDAY, THURSDAY, FRIDAY, SATURDAY, SUNDAY]
      example: MONDAY
    TimeWindow:
      type: object
      required: [dayOfWeek, startTime, endTime]
      properties:
        dayOfWeek:
          $ref: "#/components/schemas/DayOfWeek"
        startTime:
          $ref: "#/components/schemas/HoursAndMinutesTime"
        endTime:
          $ref: "#/components/schemas/HoursAndMinutesTime"
    Slot:
      type: object
      required: [dayOfWeek, startTime, endTime, maxOccupancy]
      properties:
        dayOfWeek:
          $ref: "#/components/schemas/DayOfWeek"
        startTime:
          $ref: "#/components/schemas/HoursAndMinutesTime"
        endTime:
          $ref: "#/components/schemas/HoursAndMinutesTime"
        maxOccupancy:
          type: number
          format: int32
    IanaTimeZone:
      type: string
      pattern: "^[A-Za-z_]+/[A-Za-z_]+$"
      example: "Europe/Warsaw"
    HoursAndMinutesTime:
      type: string
      pattern: "^([01]\\d|2[0-3]):[0-5]\\d$"
      example: "13:00"
    Location:
      type: object
      required: [address, longitude, latitude]
      properties:
        address:
          type: string
          example: ul. Nowowiejska 15/19, Warsaw, Poland
        longitude:
          type: number
          format: double
          example: 12.3248
        latitude:
          type: number
          format: double
          example: -67.224
...
  responses:
...
    Created:
      description: &CREATED_DESCRIPTION >
        Created resource
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            required:
              - message
            properties:
              message:
                type: string
                example: Unauthorized
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            required:
              - message
            properties:
              message:
                type: string
                example: Bad request
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            type: object
            required:
              - message
            properties:
              message:
                type: string
                example: Forbidden request
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            required:
              - message
            properties:
              message:
                type: string
                example: Internal server error
    NoContent:
      description: No content
      content:
        application/json:
          schema:
            type: object
            required:
              - message
            properties:
              message:
                type: string
                example: No content
paths:
...
  /enterprises:
    post:
      tags: [Enterprises]
      operationId: createEnterprise
      security:
        - JwtBearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/CreateEnterpriseRequest"
      responses:
        "201":
          description: *CREATED_DESCRIPTION
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateEnterpriseResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /enterprises/{enterpriseId}:
    get:
      tags: [Enterprises]
      operationId: getEnterprise
      security:
        - JwtBearerAuth: []
      parameters:
        - name: enterpriseId
          in: path
          required: true
          description: Enterprise id
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: *OK_DESCRIPTION
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetEnterpriseResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /enterprises/{enterpriseId}/employees/exists:
    get:
      tags: [EnterpriseEmployees]
      operationId: checkEmployeeUsernameExists
      security:
        - JwtBearerAuth: []
      parameters:
        - name: enterpriseId
          in: path
          required: true
          description: Enterprise id
          schema:
            type: integer
            format: int64
        - description: Username to check
          in: query
          name: username
          required: true
          schema:
            type: string
            example: jan.kowalski.1
      responses:
        "200":
          description: *OK_DESCRIPTION
          content:
            application/json:
              schema:
                type: object
                properties:
                  isExisting:
                    type: boolean
                required: [isExisting]
              examples:
                yes:
                  value:
                    isExisting: true
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /enterprises/{enterpriseId}/employees:
    get:
      tags: [EnterpriseEmployees]
      operationId: getEnterpriseEmployees
      summary: Get enterprise employees
      security:
        - JwtBearerAuth: []
      parameters:
        - name: enterpriseId
          in: path
          required: true
          description: Enterprise id
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: *OK_DESCRIPTION
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/GetEnterpriseEmployeesResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"

    post:
      tags: [EnterpriseEmployees]
      operationId: createEnterpriseEmployee
      summary: Create enterprise employee
      security:
        - JwtBearerAuth: []
      parameters:
        - name: enterpriseId
          in: path
          required: true
          description: Enterprise id
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEnterpriseEmployeeRequest"
      responses:
        "201":
          description: *CREATED_DESCRIPTION
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/CreateEnterpriseEmployeeResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /enterprises/{enterpriseId}/services:
    get:
      tags: [Enterprises]
      operationId: getEnterpriseServices
      summary: Get enterprise employees
      parameters:
        - name: enterpriseId
          in: path
          required: true
          description: Enterprise id
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: *OK_DESCRIPTION
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetEnterpriseServicesResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /enterprises/{enterpriseId}/no_custom_appointment_services:
    post:
      tags: [EnterpriseServices]
      operationId: createNoCustomAppointmentsEnterpriseService
      security:
        - JwtBearerAuth: []
      parameters:
        - name: enterpriseId
          in: path
          required: true
          description: Enterprise id
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNoCustomAppointmentsEnterpriseServiceRequest"
      responses:
        "201":
          $ref: "#/components/responses/Created"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /enterprises/{enterpriseId}/custom_appointment_services:
    post:
      tags: [EnterpriseServices]
      operationId: createCustomAppointmentsEnterpriseService
      security:
        - JwtBearerAuth: []
      parameters:
        - name: enterpriseId
          in: path
          required: true
          description: Enterprise id
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomAppointmentsEnterpriseServiceRequest"
      responses:
        "201":
          $ref: "#/components/responses/Created"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /enterprises/{enterpriseId}/patch:
    patch:
      tags: [Enterprises]
      operationId: patchEnterprise
      security:
        - JwtBearerAuth: []
      parameters:
        - name: enterpriseId
          in: path
          required: true
          description: Enterprise id
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/PatchEnterpriseRequest"
      responses:
        "200":
          description: *OK_DESCRIPTION
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /enterprises/{enterpriseId}/photos/background:
    get:
      tags: [Enterprises]
      operationId: getEnterpriseBackgroundPhoto
      parameters:
        - name: enterpriseId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: *OK_DESCRIPTION
          content:
            image/png:
              schema:
                type: string
                format: binary
          headers:
            Content-Disposition:
              schema:
                type: string
                example: inline; filename="photo.png"
        "204":
          $ref: "#/components/responses/NoContent"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /enterprises/{enterpriseId}/photos/logo:
    get:
      tags: [Enterprises]
      operationId: getEnterpriseLogoPhoto
      parameters:
        - name: enterpriseId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: *OK_DESCRIPTION
          content:
            image/png:
              schema:
                type: string
                format: binary
          headers:
            Content-Disposition:
              schema:
                type: string
                example: inline; filename="photo.png"
        "204":
          $ref: "#/components/responses/NoContent"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
...